Shortcut to Servlet
Thanks to the earnest encouragement from i2pn00b, I went and converted this thing into a servlet for supreme convenience of user interaction. (Plus that means I can put it on my eepsite!) It will lookup the b64 and b32 of any given name (or b64 I think) and if possible copy that b32 to the system clipboard.