Made a new webapp. :p
Had this in mind for a while. Basically it regularly polls to see if the 
master addressbook susidns spits out got any bigger, then sucks the 
new addresses into my very own awful database implementation that should be 
neodatis if I could be assed to figure out how to do that. Anyone visiting
the servlet page will be pleased to find (hopefully) that it keeps track
of when what hosts were added, and only sends you the hosts you haven't yet
gotten.